Move code for painting text selection cursor
