+++ crates/egui/src/text_selection/mod.rs
@@ -1,5 +1,6 @@
 //! Helpers regarding text selection for labels and text edit.
 
 mod label_text_selection;
+pub mod visuals;
 
 pub use label_text_selection::label_text_selection;
