+++ crates/egui/src/widgets/text_edit/mod.rs
@@ -9,10 +9,6 @@ mod text_buffer;
 pub mod accesskit_text;
 
 pub use {
-    builder::{paint_cursor_selection, TextEdit},
-    cursor_range::*,
-    output::TextEditOutput,
-    state::TextCursorState,
-    state::TextEditState,
-    text_buffer::TextBuffer,
+    builder::TextEdit, cursor_range::*, output::TextEditOutput, state::TextCursorState,
+    state::TextEditState, text_buffer::TextBuffer,
 };
